+++ public/ihltv.h
@@ -33,7 +33,8 @@ public:
 	virtual	IServer	*GetBaseServer( void ) = 0; // get HLTV base server interface
 	virtual	IHLTVDirector *GetDirector( void ) = 0;	// get director interface
 	virtual	int		GetHLTVSlot( void ) = 0; // return entity index-1 of HLTV in game
-	virtual float	GetOnlineTime( void ) = 0; // seconds since broadcast started
+	// dimhotepus: float -> double
+	virtual double	GetOnlineTime( void ) = 0; // seconds since broadcast started
 	virtual void	GetLocalStats( int &proxies, int &slots, int &specs ) = 0; 
 	virtual void	GetGlobalStats( int &proxies, int &slots, int &specs ) = 0;
